requirements.txt
tqdm==4.65.0
numpy==1.21.5
Pillow==9.2.0
opencv-python==4.9.0
imageio
matplotlib==3.5.2
kornia
natsort
pyyaml
plotly
lpips
open3d==0.16.0
torchmetrics
cyclonedds
pytorch-msssim
jupyter
trimesh